backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==1.4.53
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pillow>=10.3.0
aiofiles==23.2.1
pydantic==2.5.0
email-validator==2.1.0.post1
